Move Package BranchContract to common lib
