+++ yggdrash-common/src/main/java/io/yggdrash/common/store/BranchStateStore.java
@@ -18,6 +18,7 @@ package io.yggdrash.common.store;
 
 import io.yggdrash.common.Sha3Hash;
 import io.yggdrash.common.contract.vo.dpoa.ValidatorSet;
+import java.util.List;
 
 public interface BranchStateStore {
     Long getLastExecuteBlockIndex();
@@ -33,4 +34,6 @@ public interface BranchStateStore {
     void setValidators(ValidatorSet validatorSet);
 
     boolean isValidator(String address);
+
+    public List<BranchContract> getBranchContacts();
 }
