+++ yggdrash-core/src/main/java/io/yggdrash/core/blockchain/BlockChain.java
@@ -17,6 +17,7 @@
 package io.yggdrash.core.blockchain;
 
 import io.yggdrash.common.contract.ContractVersion;
+import io.yggdrash.common.store.BranchContract;
 import io.yggdrash.core.blockchain.osgi.ContractManager;
 import io.yggdrash.core.consensus.ConsensusBlock;
 import io.yggdrash.core.consensus.ConsensusBlockChain;
