+++ yggdrash-core/src/main/java/io/yggdrash/core/blockchain/Branch.java
@@ -21,6 +21,7 @@ import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
 import io.yggdrash.common.config.Constants;
 import io.yggdrash.common.crypto.HexUtil;
+import io.yggdrash.common.store.BranchContract;
 import io.yggdrash.common.utils.FileUtil;
 import io.yggdrash.common.utils.JsonUtil;
 import org.apache.commons.io.IOUtils;
