+++ yggdrash-core/src/main/java/io/yggdrash/core/blockchain/BranchBuilder.java
@@ -18,6 +18,7 @@ package io.yggdrash.core.blockchain;
 
 import com.google.gson.JsonArray;
 import com.google.gson.JsonObject;
+import io.yggdrash.common.store.BranchContract;
 
 import java.util.ArrayList;
 import java.util.List;
