+++ yggdrash-core/src/main/java/io/yggdrash/core/blockchain/PrepareBlockchain.java
@@ -17,6 +17,7 @@
 package io.yggdrash.core.blockchain;
 
 import io.yggdrash.common.contract.ContractVersion;
+import io.yggdrash.common.store.BranchContract;
 import org.apache.commons.io.IOUtils;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
