+++ yggdrash-core/src/test/java/io/yggdrash/core/blockchain/BranchContractTest.java
@@ -18,6 +18,7 @@ package io.yggdrash.core.blockchain;
 
 import com.google.gson.JsonObject;
 import com.google.gson.JsonParser;
+import io.yggdrash.common.store.BranchContract;
 import org.junit.Assert;
 import org.junit.Test;
 import org.slf4j.Logger;
