+++ yggdrash-core/src/test/java/io/yggdrash/core/blockchain/osgi/ContractChannelCouplerTest.java
@@ -42,7 +42,12 @@ public class ContractChannelCouplerTest {
 
         t.callContractChannelQuery("TEST1","yesmanQuery");
 
+    }
+
+    @Test
+    public void callChnnelMethodTest() {
 
     }
+
 
 }
